requirements.txt
keras==2.7.0
numpy==1.21.5
Pillow==9.0.0
streamlit==1.4.0
tensorflow==2.7.0
